fix: 修复NumberSource实体fee字段类型为Double，同步更新所有引用处
- NumberSource.fee: Integer → Double (匹配数据库DECIMAL(10,2))
- 更新DoctorController、DoctorWorkController、AppointmentController中的fee赋值
- 更新ScheduleInitializer、ScheduleAutoTask中的fee赋值
- 添加parseDouble工具方法到DoctorWorkController

+++ src/main/java/com/appointment/entity/NumberSource.java
@@ -7,7 +7,7 @@ public class NumberSource {
     private Integer scheduleId;
     private Integer totalNum;
     private Integer remainNum;
-    private Integer fee;
+    private Double fee;
     private Integer status;
     private Date createTime;
     private Date updateTime;
@@ -44,11 +44,11 @@ public class NumberSource {
         this.remainNum = remainNum;
     }
 
-    public Integer getFee() {
+    public Double getFee() {
         return fee;
     }
 
-    public void setFee(Integer fee) {
+    public void setFee(Double fee) {
         this.fee = fee;
     }
